python/pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kora-sdk"
version = "1.1.0"
description = "Python SDK for the Kora authorization engine — deterministic spend authorization for AI agents"
readme = "README.md"
license = {text = "AGPL-3.0-or-later"}
requires-python = ">=3.10"
authors = [
    {name = "Kora Protocol"},
]
keywords = ["kora", "authorization", "ai-agents", "spending", "ed25519", "deterministic", "fintech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Security",
    "Topic :: Software Development :: Libraries",
]
dependencies = [
    "pynacl>=1.5.0",
    "requests>=2.28.0",
]

[project.urls]
Homepage = "https://github.com/kora-protocol/kora"
Repository = "https://github.com/kora-protocol/kora"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
